+++ tests/test_subcommand_taxtable.py
@@ -3,6 +3,7 @@
 from os import path
 import logging
 
+from cStringIO import StringIO
 from sqlalchemy import create_engine
 
 from . import config
@@ -20,6 +21,7 @@ echo = False
 
 dbname = config.ncbi_master_db
 
+
 class TaxTableSetup(TestBase):
 
     def setUp(self):
@@ -29,6 +31,7 @@ class TaxTableSetup(TestBase):
     def tearDown(self):
         self.engine.dispose()
 
+
 class TestTaxonomyInit(TaxTableSetup):
 
     def test01(self):
@@ -37,14 +40,15 @@ class TestTaxonomyInit(TaxTableSetup):
     def test02(self):
         self.assertRaises(KeyError, self.tax._node, 'buh')
 
+
 class TestGetLineagePrivate(TaxTableSetup):
 
     def test01(self):
         lineage = self.tax._get_lineage('1')
-        self.assertTrue(lineage == [('root','1')])
+        self.assertTrue(lineage == [('root', '1')])
 
     def test02(self):
-        tax_id = '1280' # staph aureus
+        tax_id = '1280'  # staph aureus
 
         self.assertFalse(tax_id in self.tax.cached)
         lineage = self.tax._get_lineage(tax_id)
@@ -53,11 +57,12 @@ class TestGetLineagePrivate(TaxTableSetup):
         self.assertTrue(lineage[-1][0] == 'species')
 
     def test03(self):
-        tax_id = '30630' # deprecated; Microtus levis Taxonomy ID: 537919
+        tax_id = '30630'  # deprecated; Microtus levis Taxonomy ID: 537919
 
         self.assertFalse(tax_id in self.tax.cached)
         self.assertRaises(KeyError, self.tax._get_lineage, tax_id,
-                merge_obsolete=False)
+                          merge_obsolete=False)
+
 
 class TestGetMerged(TaxTableSetup):
 
@@ -67,9 +72,10 @@ class TestGetMerged(TaxTableSetup):
         self.assertEqual(tax_id, merged)
 
     def test02(self):
-        tax_id = '30630' # deprecated; Microtus levis Taxonomy ID: 537919
+        tax_id = '30630'  # deprecated; Microtus levis Taxonomy ID: 537919
         merged = self.tax._get_merged(tax_id)
         self.assertFalse(merged is None)
+
 
 class TestTaxNameSearch(TaxTableSetup):
 
@@ -82,7 +88,8 @@ class TestTaxNameSearch(TaxTableSetup):
         self.assertRaises(KeyError, self.tax.primary_from_name, 'buggabugga')
 
     def test03(self):
-        tax_id, tax_name, is_primary = self.tax.primary_from_name('Gemella Berger 1960')
+        tax_id, tax_name, is_primary = self.tax.primary_from_name(
+            'Gemella Berger 1960')
 
         self.assertTrue(tax_id == '1378')
         self.assertFalse(is_primary)
@@ -97,7 +104,6 @@ class TestSynonyms(TaxTableSetup):
         self.tax.synonyms(tax_name='Gemella')
 
 
-
 class TestGetLineagePublic(TaxTableSetup):
 
     def test01(self):
@@ -107,7 +113,7 @@ class TestGetLineagePublic(TaxTableSetup):
         self.assertTrue(lineage['rank'] == 'root')
 
     def test02(self):
-        tax_id = '1280' # staph aureus
+        tax_id = '1280'  # staph aureus
 
         self.assertFalse(tax_id in self.tax.cached)
         lineage = self.tax.lineage(tax_id)
@@ -116,25 +122,29 @@ class TestGetLineagePublic(TaxTableSetup):
 
         keys = set(lineage.keys())
         ranks = set(self.tax.ranks)
-        self.assertTrue(keys - ranks == set(['parent_id', 'tax_id', 'rank', 'tax_name']))
+        self.assertTrue(
+            keys - ranks == set(['parent_id', 'tax_id', 'rank', 'tax_name']))
 
     def test03(self):
-        tax_id = '1378' # Gemella; lineage has two successive no_rank taxa
+        tax_id = '1378'  # Gemella; lineage has two successive no_rank taxa
         lineage = self.tax.lineage(tax_id)
         self.assertTrue(lineage['rank'] == 'genus')
 
         keys = set(lineage.keys())
         ranks = set(self.tax.ranks)
-        self.assertTrue(keys - ranks == set(['parent_id', 'tax_id', 'rank', 'tax_name']))
+        self.assertTrue(
+            keys - ranks == set(['parent_id', 'tax_id', 'rank', 'tax_name']))
 
     def test04(self):
-        self.assertRaises(ValueError, self.tax.lineage, tax_id=None, tax_name=None)
+        self.assertRaises(
+            ValueError, self.tax.lineage, tax_id=None, tax_name=None)
 
     def test05(self):
-        self.assertRaises(ValueError, self.tax.lineage, tax_id='1', tax_name='root')
+        self.assertRaises(
+            ValueError, self.tax.lineage, tax_id='1', tax_name='root')
 
     def test06(self):
-        tax_id = '1378' # Gemella; lineage has two successive no_rank taxa
+        tax_id = '1378'  # Gemella; lineage has two successive no_rank taxa
         tax_name = 'Gemella'
         self.tax.lineage(tax_name=tax_name)
 
@@ -142,7 +152,7 @@ class TestGetLineagePublic(TaxTableSetup):
         # self.assertTrue(lineage['rank'] == 'genus')
 
     def test07(self):
-        tax_id = '30630' # deprecated; Microtus levis Taxonomy ID: 537919
+        tax_id = '30630'  # deprecated; Microtus levis Taxonomy ID: 537919
         self.tax.lineage(tax_id=tax_id)
 
 
@@ -157,10 +167,11 @@ class TestMethods(TaxTableSetup):
 
     # Commented: Too varying
     # def test03(self):
-    #     res = self.tax.add_source(name='new source', description='really new!')
-    #     res = self.tax.add_source(name='new source', description='really new!')
+    #     res = self.tax.add_source(name='new source',
+    #                               description='really new!')
+    #     res = self.tax.add_source(name='new source',
+    #                               description='really new!')
     #     self.assertEqual(res, (2, False))
-
 
     # def test04(self):
     #     self.tax.add_node(tax_id = "186802_1",
@@ -171,6 +182,7 @@ class TestMethods(TaxTableSetup):
 
 
 class TestWriteTable(TaxTableSetup):
+
     """
     test tax.write_table - note that this method produces output.
     """
@@ -186,14 +198,24 @@ class TestWriteTable(TaxTableSetup):
         self.assertTrue(path.isfile(self.fname))
 
     def test02(self):
-        tax_id = '1280' # staph aureus
+        tax_id = '1280'  # staph aureus
         self.tax.lineage(tax_id)
 
     def test03(self):
-        tax_id = '1378' # Gemella; lineage has two successive no_rank taxa
+        tax_id = '1378'  # Gemella; lineage has two successive no_rank taxa
         self.tax.lineage(tax_id)
 
     def test04(self):
-        tax_id = '1378' # Gemella; lineage has two successive no_rank taxa
-        for tax_id in ['1378','1280','131110']:
+        tax_id = '1378'  # Gemella; lineage has two successive no_rank taxa
+        for tax_id in ['1378', '1280', '131110']:
             self.tax.lineage(tax_id)
+
+    def test05(self):
+        """
+        Do all ranks appear with full=True
+        """
+
+        outstr = StringIO()
+        self.tax.write_table(taxa=None, csvfile=outstr, full=True)
+        columns = outstr.getvalue().rstrip().strip('"').split('","')
+        self.assertFalse(set(taxtastic.ncbi.ranks) - set(columns))  # empty set