+++ CHANGES.rst
@@ -5,9 +5,11 @@
 0.5.4-master
 =====
 
- * Add ``taxit taxtable --full`` for outputing all ranks in header.
+ * Added ``taxit taxtable --full`` for outputing all ranks in header plus test case
  * ``taxit taxids`` can process accession numbers by querying NCBI
  * ``taxit taxids`` output now in csv format query,tax_id
+ * ``taxit taxids --no-children`` return only immediate taxid ignoring all children taxids
+ * Added some taxit taxids unittests in test_subcommands.py
 
 0.5.3
 =====